LoRaDB-manager/requirements.txt
textual>=0.47.0
docker>=7.0.0
pyyaml>=6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.0
PyJWT>=2.8.0
pyperclip>=1.8.2
